+++ internal/resource/config/passwordcredentialvalidator/password_credential_validator_data_source.go
@@ -1,0 +1,145 @@
+package passwordcredentialvalidator
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	client "github.com/pingidentity/pingfederate-go-client/v1125/configurationapi"
+	"github.com/pingidentity/terraform-provider-pingfederate/internal/datasource/common/id"
+	"github.com/pingidentity/terraform-provider-pingfederate/internal/datasource/common/pluginconfiguration"
+	"github.com/pingidentity/terraform-provider-pingfederate/internal/datasource/common/resourcelink"
+	"github.com/pingidentity/terraform-provider-pingfederate/internal/resource/config"
+	internaltypes "github.com/pingidentity/terraform-provider-pingfederate/internal/types"
+)
+
+// Ensure the implementation satisfies the expected interfaces.
+var (
+	_ datasource.DataSource              = &passwordCredentialValidatorDataSource{}
+	_ datasource.DataSourceWithConfigure = &passwordCredentialValidatorDataSource{}
+)
+
+// PasswordCredentialValidatorDataSource is a helper function to simplify the provider implementation.
+func PasswordCredentialValidatorDataSource() datasource.DataSource {
+	return &passwordCredentialValidatorDataSource{}
+}
+
+// passwordCredentialValidatorDataSource is the datasource implementation.
+type passwordCredentialValidatorDataSource struct {
+	providerConfig internaltypes.ProviderConfiguration
+	apiClient      *client.APIClient
+}
+
+// GetSchema defines the schema for the datasource.
+func (r *passwordCredentialValidatorDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
+	schema := schema.Schema{
+		Description: "Describes a password credential validator plugin instance.",
+		Attributes: map[string]schema.Attribute{
+			"name": schema.StringAttribute{
+				Description: "The plugin instance name. The name can be modified once the instance is created. Note: Ignored when specifying a connection's adapter override.",
+				Optional:    false,
+				Computed:    true,
+			},
+			"plugin_descriptor_ref": schema.SingleNestedAttribute{
+				Description: "Reference to the plugin descriptor for this instance. The plugin descriptor cannot be modified once the instance is created. Note: Ignored when specifying a connection's adapter override.",
+				Optional:    false,
+				Computed:    true,
+				Attributes:  resourcelink.ToDataSourceSchema(),
+			},
+			"parent_ref": schema.SingleNestedAttribute{
+				Description: "The reference to this plugin's parent instance. The parent reference is only accepted if the plugin type supports parent instances. Note: This parent reference is required if this plugin instance is used as an overriding plugin (e.g. connection adapter overrides)",
+				Optional:    false,
+				Computed:    true,
+				Attributes:  resourcelink.ToDataSourceSchema(),
+			},
+			"configuration": pluginconfiguration.ToDataSourceSchema(),
+			"attribute_contract": schema.SingleNestedAttribute{
+				Description: "The list of attributes that the password credential validator provides.",
+				Optional:    false,
+				Computed:    true,
+				Attributes: map[string]schema.Attribute{
+					"core_attributes": schema.ListNestedAttribute{
+						Description: "A list of read-only attributes that are automatically populated by the password credential validator descriptor.",
+						Optional:    false,
+						Computed:    true,
+						NestedObject: schema.NestedAttributeObject{
+							Attributes: map[string]schema.Attribute{
+								"name": schema.StringAttribute{
+									Description: "The name of this attribute.",
+									Optional:    false,
+									Computed:    true,
+								},
+							},
+						},
+					},
+					"extended_attributes": schema.ListNestedAttribute{
+						Description: "A list of additional attributes that can be returned by the password credential validator. The extended attributes are only used if the adapter supports them.",
+						Optional:    false,
+						Computed:    true,
+						NestedObject: schema.NestedAttributeObject{
+							Attributes: map[string]schema.Attribute{
+								"name": schema.StringAttribute{
+									Description: "The name of this attribute.",
+									Optional:    false,
+									Computed:    true,
+								},
+							},
+						},
+					},
+					"inherited": schema.BoolAttribute{
+						Description: "Whether this attribute contract is inherited from its parent instance. If true, the rest of the properties in this model become read-only. The default value is false.",
+						Optional:    false,
+						Computed:    true,
+					},
+				},
+			},
+		},
+	}
+
+	id.ToSchema(&schema)
+	id.ToDataSourceSchemaCustomId(&schema,
+		"validator_id",
+		true,
+		false,
+		"The ID of the plugin instance. The ID cannot be modified once the instance is created. Note: Ignored when specifying a connection's adapter override.")
+	resp.Schema = schema
+}
+
+// Metadata returns the datasource type name.
+func (r *passwordCredentialValidatorDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_password_credential_validator"
+}
+
+func (r *passwordCredentialValidatorDataSource) Configure(_ context.Context, req datasource.ConfigureRequest, _ *datasource.ConfigureResponse) {
+	if req.ProviderData == nil {
+		return
+	}
+
+	providerCfg := req.ProviderData.(internaltypes.ResourceConfiguration)
+	r.providerConfig = providerCfg.ProviderConfig
+	r.apiClient = providerCfg.ApiClient
+
+}
+
+func (r *passwordCredentialValidatorDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
+	var state passwordCredentialValidatorModel
+
+	diags := req.Config.Get(ctx, &state)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+	apiReadPasswordCredentialValidators, httpResp, err := r.apiClient.PasswordCredentialValidatorsAPI.GetPasswordCredentialValidator(config.ProviderBasicAuthContext(ctx, r.providerConfig), state.ValidatorId.ValueString()).Execute()
+	if err != nil {
+		config.ReportHttpError(ctx, &resp.Diagnostics, "An error occurred while getting a Password Credential Validator", err, httpResp)
+		return
+	}
+
+	// Read the response into the state
+	diags = readPasswordCredentialValidatorResponse(ctx, apiReadPasswordCredentialValidators, &state, state.Configuration, false)
+	resp.Diagnostics.Append(diags...)
+
+	// Set refreshed state
+	diags = resp.State.Set(ctx, &state)
+	resp.Diagnostics.Append(diags...)
+}